Renamed file and fixed bug with udp_ip in mchswork.
