+++ caen_tools/SystemCheck/scripts/structures.py
